pyproject.toml
[project]
name = "agent-harness"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = ["litellm", "pyyaml"]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio"]
grpc = ["grpcio>=1.60.0", "grpcio-tools>=1.60.0", "protobuf>=4.25.0"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
